bugfix: fixed a segfault introduced in c04bc97 when testing an nginx configuration containing 'lua_shared_dict' directive(s).
In commit c04bc97, we avoided running `init_by_lua*` in signaller processes and
when testing the Nginx configuration. However, when the Nginx configuration
contains `lua_shared_dict` directives, the execution of `init_by_lua*` will be
post-poned when testing the configuration.

This case was not handled and led to the segmentation fault reported in
openresty/lua-nginx-module#1462.

Co-authored-by: spacewander <spacewanderlzx@gmail.com>