chore: makefile: allowed .tt2 templates to be compiled from a subsystem directory.
This is useful when a given file has been converted to a .tt2 template,
but not ready to be ported to another subsystem yet.

E.g. the ngx.pipe API is now ported to a template (and compiles with the
ngx_stream_lua module), but tests are not ported yet. In order to not
revert the template, we make it part of the http subsystem only until
tests are ported to the ngx_stream_lua module.