bugfix: ignore the `raw` argument when acquiring UDP request socket.
In the documentation, we stated that the `raw` argument of the
`ngx.req.socket` API is always ignored in the stream subsystem and we
does so for TCP request sockets. However, that's not the case for UDP
downstream socket. If the user specifies the `raw` argument manually a
LuaJIT assertion error like this would occur:

```
lj_api.c:1000: lua_rawseti: Assertion `(((uint32_t)((index2adr(L, idx))->it64 >> 47)) == (~11u))' failed.
```

Here are the crash dumps:

```
(gdb) bt
    at ../ngx_stream_lua-0.0.8/src/ngx_stream_lua_util.c:835
```

this PR makes sure the `raw` argument is properly ignored as we have
stated in the documentation and maintains the same behavior between TCP
and UDP sessions.