change: changed from graceful shutdown to abort when lua_resume return LUA_ERRMEM.
If the worker voluntary enter into graceful shutdown state, master will not respawn worker with the same worker id until the worker exit.
when reuseport is on, the new connections assigined to the voluntary graceful shutdown worker will not be accepted.
when reuseport is off, the new connections will be accepted by other processes which are still working. But if all workers are in voluntary shutdown state, the new connections will not be accepted.

So the worker should not enter into the voluntary graceful shutdown state when LUA_ERRMEM occur.